requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.2.2
Django==2.0.7
django-filter==2.0.0
django-js-asset==1.1.0
django-model-utils==3.1.2
django-mptt==0.9.0
django-notifications-hq==1.5.0
django-pure-pagination==0.3.0
djangorestframework==3.8.2
idna==2.7
jsonfield==2.0.2
Markdown==2.6.11
mistune==0.8.3
Pillow==5.2.0
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
